done cart

+++ Unleashed/OrderService/DTOs/ResponesDtos/VariationResponseDto.cs
@@ -4,6 +4,6 @@
     {
         public int VariationId { get; set; }
         public Guid ProductId { get; set; }
-        public decimal VariationPrice { get; set; }
+        public double VariationPrice { get; set; }
     }
 }
